Add dummy binning parameters to PSP102
PSP102 already defined these parameters, but only if the "Binning"
macro was set, which also makes other changes.  As they are only dummy
parameters (and acknowledged as such in their comments), I'm copying
the definitions to the "else" part of the Binning ifdef, too.

The PSP102 is regenerated into C++ by an experimental version of
Xyce/ADMS, not the one on the master branch.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1198