remove getMaxTimeStepSize function from resistor3
Most devices should not implement this function, and the implemention in
this particular device just returned a very large number (but not the
official defaultMaxTimeStep).   So, it had no impact on actual runtime,
except to be a wasted function call.

I've been running profilers a lot today and discovered that for some
circuits (the RLC transmission line in particular) the
"getMaxTimeStepSize" evaluation in the device package is surprisingly
expensive.  So, I am trying to fix that and as part of that I checked
which devices actually have this function implemented.  The answer is
not very many.