optimize DeviceMgr::getMaxTimeStepSize function
This function is called by the time integrator at the end of every
successful time step.  Most devices do not have a maximum time step to
impose, so this function should be very cheap.  However, it is not.

I recently ran a profiler on the RLC ladder regression test.  This is a
test that takes several minutes to run.  I only sampled it for about 20
seconds in the middle of the transient.  Of those 20 seconds I saw the
following:

About 15 seconds was for the function
"Transient::takeAnIntegrationStepSize".  As this function contains most of
the numerical work for a time step, it makes sense that it would be
expensive.  So I didn't dig into it (for now).

The other 5 seconds was taken up by
"Transient::doProcessSuccessfulStep".  This was unexpected, as this
function should mostly be bookkeeping.  Within this 5 seconds, about
2.86 seconds was taken up by
"TimeInteg::StepErrorControl::updateMaxTimeStep".  Almost *all* of this
(2.85s) was taken up by "DeviceMgr::getMaxTimeStepSize".

Of this 2.85s, nearly 1s was taken up by a dynamic_cast that wasn't
necessary.   The function was also expensive because it called
"getMaxTimeStep" on every single device instance, even though very few
of them had implemented that function.

I've rewritten the function to (1) only call devices that have
implemented the getMaxTimeStepSize function, and (2) don't do any
dynamic casting.  This should fix this particular inefficiency.