add fix to speedup Expr::set_sim_time/freq
In the recent work on the SAR_ADC circuit, Ting made the observation,
based on profiler results, that a lot of time was being spent in the
expression library function "set_sim_time".  This seems strange as the
function would not appear to be very complicated.

It turns out that it is called a lot as part of the breakpoints function
in the expression library.  I didn't try to address that directly but I
did attempt to make the function itself faster.

So, pursuant to that effort, this commit does the following:

- Replaces the manually implemented serach on a std::vector of
std:strings with a std::find based search.  I'm guessing this won't
matter much, but I figured it wouldn't hurt

- Modifies the code so that this search only needs to be performed a
small number of times.  Once the expression is established, the index
for the "TIME" variable should not change.  So, it is silly to do a
search to find it every single time.   I initially tried to make it so
that it would only search for it once, but it was difficult to get that
to work.  So the solution I settled on looks for the index a few times
before either giving up or finding the final value.

This was actually already being done in one of the "breaks" related
functions so I did something very similar.

While I was at it, I did the same change for the "FREQ" special
variable.  It gets set a lot as well, even in transient(?) runs.

My shortened ADC test case is around a minute faster (out of about
7 min) as the result of these changes.