Fix for Error Checking in C Device Constructor
This commit fixes the error checking for "unsupported soluton
dependent parameters" in the constructor for the C device.  That
error check used the number of variables in the parameter's
expression.  However that quantity was not being set to a value,
in the function DeviceEntity::setDependentParameter(), when the
parameter (e.g., L for the C device) is not allowed to be solution
dependent.

SON Bug 860
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=860