Revert "Fix for Error Checking in C Device Constructor"
This reverts commit fe2991e4def91b641a192ec4bdeea77935a9de16.

The reason for this reversion is that in order for a device parameter
to be solution dependent, one has to do several things: 1) change its
"setExpressionAccess" setting to type SOLN_DEP; and 2) add some code
manually into N_IO_DistToolBase.C.  The DeviceEntity class then
has the ability to check whether that parameter has solution dependence,
and whether its device has designated it as legal for that parameter
to be solution dependent.  So, this particular error check in the
C-device constructor was deemed "redundant" and will be removed in
the next commit.