Fix a bug in S-parameter analysis
The S-parameter analysis can segfault if no port device
is specified and sparcal=1. The port device is treated as a voltage
source internally. In a netlist, the circuit can have both the
port devices and voltage sources. If no port device is specified and
voltage sources are specified, it can lead to different number of
B matrix entries and number of ports. This can lead to segfault.
This commit fixes this