Add variables to BSIM4 member initialization lists
When adding a bunch of "given" boolean variables to the BSIM4 device
last week, which I did to speed up the "processParams" function, I
noticed that there were a large number of variables in the BSIM4 source
code that were not being initialized in the contructor member
initialization lists.    I made a note at the time to fix this, and this
commit does that.

The main reason I wanted to fix it, is that I will probably eventually
go back to the BSIM4 and add more "given" variables to it, and I was
going to feel compelled to fix these while working on it.  But these
aren't very important changes.  They only come up if you compile with
-Weffc++, which I don't recommend doing most of the time.

This commit includes a few small member initialization list corrections
in 2 other classes as well.

As these are very minor changes and for future "initializaiton list"
fixes I doubt I'll write more than the one-line summary.