Fix bitwise pointer copy in SweepParam
Discovered this while compiling the BSIM4 with lots of warnings enabled.
One bit of guidance from Scott Meyers Effective C++ books is that if a
class has raw pointer members, they need to be dealt with in a manually
constructed copy constructor and also the operator= function.

If they aren't, then any copies made of the class will include bitwise
pointer copies, which is usually undesirable.

Another bit of guidance he gives is that if you can handle something
without involving pointers that's a good thing to do.

I did not follow that guidance in the SweepParam class, when I hacked it
to handle sampling.  This only matters in the C++11 version of the
class, which was using pointers to objects such as
std::uniform_real_distribution<double>, from the <random> library.
It wasn't necessary to handle those objects with pointers, or even have
them directly in the SweepParam class at all.

The SweepParams objects are stored in a std::vector, so
the copy constructor and/or operator= function are used,  and so
copies of the raw pointers were being made. This didn't appear
(superficially at least) to break the code, even in valgrind.  But it
was the wrong thing to do, so I've re-written the code so that the
SweepParam object no longer owns any kind of ramdom distribution object
(pointer or otherwise).

Partly I worked thru this just to get back into the UQ source code,
preparing for another UQ methods "sprint".