Remove Redundant Error Check in C-Device Constructor
This commit removes a redundant error check in the C-device
constructor, that did not work consistently anyway.  It is
redundant because the DeviceEntity class already checks whether
a parameter is allowed to be "solution dependent", and emits
an error if it is not.  The C and Q parameters are the only C device
parameters that have setExpressionAccess(ParameterType::SOLN_DEP).

SON Bug 1197
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1197