Update version to 6.11.1
In this abbreviated release "cycle", we're just cherry-picking a few
commits from develop to release branches, and calling it a new
release.  I am therefore NOT doing the whole
"development/stable/QA/production" lifecycle for this release, as it
is only a tiny delta off of the 6.11 release, which went through that
whole lifecycle.  We're going directly from 6.11.0 PROD to 6.11.1 PROD.