Refactor Parsing of .DATA Lines to Improve Error Handling
This commit refactors how .DATA lines are parsed, and then used
to create sweep vectors for .AC, .DC, .NOISE and .STEP lines. The
root problem was that the function convertData() was being called
unconditionally for all .AC, .DC, .NOISE and .STEP lines.  That led
to cases where invalid combinations of those lines with .DATA lines
would cause "inscrutable" error messages. A summary of the changes is:

  1) convertData() and the corresponding convertDataToSweepParams()
functions are now of type bool.

  2) a common function isDataSpecified() is now used to determine
if the .AC, .DC, .NOISE and .STEP lines are using DATA=<name>. That
function can set the dataSpecification_ variable which is in the
AnalysisBase class now.

This allows a more common design pattern for only converting the
DATA=<name> request into SweepParams, in the various doInit()
functions, if DATA=<name> is actually used on the .AC, .DC, .NOISE
or .STEP line.  This improves error handling.  It also allows the
same approach to be used for -remeasure for .DC.

Finally, this refactor preserves the ability to have multiple .DC
or .STEP lines in the netlist.  For .AC and .NOISE lines, the last
such line is still the only one used.

SON Bug 1188
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1188