Remove ifdef and tests for NOX sizetype
In Trilinos 11.14, NOX changed the return type of the length() method of
its NOX abstract vector class to NOX::size_type. This resulted in the
introduction of Xyce_NOX_SIZETYPE in an ifdef, and tests for the return
type in the CMake and autotools. See commit 946be9686 for details.

Since then, we've made Trilinos 12.4.2 the minimum we support (and I'm
not even sure if that's accurate). Having that unnecessary ifdef is
causing minor issues with the CMake refactor. It's time for it to go.