Remove Xyce_NOX_USE_VECTOR_COPY ifdef & config
If Xyce is compiled without Xyce_NOX_USE_VECTOR_COPY defined, then the
build fails in that section. (Until now this could be done using a
configuration option. Obviously that hasn't happend in a while since
that build configuration is broken.) After chatting with Heidi, we
decided to just remove the option, and if it needs improvement later
(which it likely will, since NOX vector copies are expensive), that can
can be done without the overhead of the ifdef.

Also, the ifdef has been causing small irriation with the CMake
refactor.