Initial Commit for Supporting .DATA with .NOISE
This commit does two things.  First, it provides the initial support
for the use of .DATA with .NOISE.  It is incomplete, since it does
not calculate the totalOutputNoise_ and totalInputNoise_ integrals
yet.  So, those values (which would be NaNs) are not output to
stdout, if DATA=<name> is used on the .NOISE line.  The issue is
that those integrals are not valid unless the frequency values in
the .DATA statement are monotonically increasing.

This commit also adds the error handling that was done for SON
Bug 1146 for .AC lines, to .NOISE lines.  This improves error
handling when DATA=<name> is not used on the .NOISE lines.

SON Bug 1042
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1042