Add constructor and initializers to Depend struct
The "Depend" struct is used to hold information about dependent params
in the device package.  It was written as a C-style struct, and its
members are not consistently initialized by the code that instantiates
the objects.

Recent work on the error handling of the solution dependent capacitor
has exposed places where uninitialized member variables of this struct
are being used in conditionals when there is an error in the
netlist (i.e, the execution has taken a code path that skips over
proper set-up of a Depend struct).  This appears to be the reason for
the instability of the CAPACITOR/ErrorMessageTest test case in recent
days.

This commit adds a constructor to the Depend class (struct) that
initializes the int members of the class to safe values.  I believe
this will cure the instability of that test.