Calculate Total Noise Int. for .DATA with .NOISE
This commit adds the capability to calculate the integrals of
the total input noise and the total output noise when DATA=<name>
is used on a .NOISE line.  The frequencies given in the DATA
table must be monotonically increasing.  If not then a warning
message is emitted during netlist parsing, and the calculation
of those two integrals (and their output to stdout) is suppressed.

SON Bug 1042
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1042