Fix a typo in std PCE outputs
For some of the std outputs in this file, I mistakenly used "samping"
instead of "sampling".

This same typo is in the shell scripts associated with the regression
tests of this capability, so they will need updating as well.  This
probably explains why it took so long for me to get this test to work a
few months ago; I was spelling it correctly in the shell script and so
it was failing.