Updated to use the new SweepParam use of <random>
On the development branch I recently modified the SweepParam class so
that it no longer owned pointers to <random> distribution objects.  This
was done partly to make the code cleaner, and also to make certain
compiler warnings to go away.

This commit updates the new intrusive PCE analysis class (which is not
present on the development branch) to be consistent with that change.