remove pointless capacitor calculations from bsim4
This commit is motivated by my attempts to optimize Xyce simulations of
the SAR ADC circuit from USC (for the ERI Summit).  This is not a big
optimization, but it did come up in my profiling, and it was easy to do.

It removes variables and calculations from the
MOSFET_B4::setupCapacitors_oldDAE function.  A little history is in
order to explain this change.

The BSIM4 (implemented by Tom in roughly 2007) follows a lot of the same
code patterns of the BSIM3 (implemented by me in roughly 2000).

The BSIM3 model, like all early models of Xyce, was originally
implemented for the "old DAE" (or really ODE) formulation.  This
basically matched what spice3 did, and did not have a separate f and q
terms.   Later, we implemented "new DAE", which meant models with
separated f and q.  For a very long time, many models supported both.

Hence, in the BSIM3/4 there is a setupCapacitors_oldDAE function, and also
a setupCapacitors_newDAE function.  The second one is the one that
really matters, and that has been true for years.  In most devices,
anything related to old-DAE has been removed at this point, but there
are a few exceptions.

The BSIM3 (and BSIM4) are among those exceptions. The BSIM3/4 was a
complex enough device that a few things in the oldDAE capacitor
function were still needed, so it was never removed.  It *should* have
been refactored so that only the bare essentials were still there, and
the superfluous stuff (mostly capacitor conductances, which we no longer
use) were removed.  But that refactor never happened.

As I've been working to speed Xyce up on the SAR ADC circuit, I figured
I might as well take a crack at this.

This commit does NOT yet do the equivalent change in the BSIM3, but that
should also get done eventually.