make DeviceMgr::updateDependentParameters readable
I've been working to optimize Xyce simulations for a moderate sized
circuit.  One thing I observed was that the expression-dependent
parameters seemed to be getting updated more frequently than they should
be.  The function from which this all happens is
DeviceMgr::updateDependentParameters_.

I thought that I might need to hack on this function a bit to fix the
issue.  That may still be the case, but I've discovered other functions
(particularly in the expression library and also the BSIM4) that are a
better target.

However, in order for me to better understand this function, I decided
to change some of the local variable names to be compliant with our code
standard, and also be more human readable and searchable.  Variable
names such as "g", or "g_i" should avoided.  This commit makes some
changes like this.  It does not functionally change the function.