Allow .PRINT AC Line to Output RF Param Ops
This commit adds the capability to output individual S-, Y- and
Z-parameter values via the .PRINT AC line.  So, the following
operators are now legal: S, SR, SI, SM, SP, SDB, Y, YR, YI, YM, YP,
YDB, Z, ZR, ZI, ZM, ZP and ZDB.  Highlights of the commit are as
follows:

  1) The type RFparamsData was defined as a map of pointers to the
Teuchos matrices of S-, Y- and Z-parameters generated by a .LIN analysis
in the AC object.  This approach allowed 6 operator classes to cover
S, Y and Z parameters, with the small inefficiency of the Operator
classes having to index into a map in their get functions.  The map key
is the first letter (S, Y and Z) of the operator name.  The various
AC-related outputter functions were modified to use the RFparamsData type.

  2) The SParams, YParam and ZParam matrices are all made by the AC
object when a .LIN analysis is done.  These matrices should be small
(e.g, of the size of the number of ports), and this approach means
that the Y-to-S and Y-to-Z conversions are done only once in the AC
object.

  3) The SP, YP and ZP operators output in radians, not degrees.

  4) The error checking in struct RFparamsVariableOpBuilder only
checks for:1) a request for S, Y or Z operators where the analysis type
is not AC, or 2) if the operator indices (e.g, S(0,0)) are < 1.  Error
checking for operator indices > number-of-ports is not implemented yet.
Xyce will output "all zeros" in that case.

  5) The changes to printLineDiagnostics() had to account for
the fact that "SENS" is a valid parameter tag.

  6) The expansion of S() into SR() and SI() operators, on a
.PRINT AC line, occurs in createOps() in N_IO_Outputter.C.

  7) The changes to N_UTL_Op.h were needed so that an error
message for say S(-1,-1) generates the correct message, instead
of three confusing error messages for S, -1 and -1.

  8) The changes to OutputMgr::prepareOutput() were needed so
that the -o command line option only makes the Touchstone output
if a .LIN analyis is done.  Otherwise, the AC output occurs for
-o output.

SON Bug 1199
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1199