more work in setting up embedded PCE class
This class originated as a renamed EmbeddedSampling class as a starting
point.

This commit removes all the code associated with regression PCE, as this
class isn't going to do that kind of PCE.  It also removes the boolean
(and related if-statements) for setting projection PCE, as this type of
analysis will always do projection.

Finally, I also did some very minor cleanup.