remove several unneccessary regression PCE objects
As noted in a previous commit, this analysis class only does projection
style PCE calculations.  It doesn't do regression PCE, unlike the
EmbeddedSamling class.

But, this class was created by initially copying the embedded sampling
class and then iteratively modifying it.   I previously removed a bunch
of regression PCE code, but forgot to remove some of the variables from
the header file.