fix hackEnsembleOutput function for converToPrn.py
When outputting tecplot format files, this function would output, on the
first line:

TITLE="the title", VARIABLES=

... with no spaces between TITLE, =, or the first quotes.   As it turns
out, this confuses the convertToPrn.py program in the test suite, and
causes the conversion to fail.  Probably the right thing to do is fix
the convertToPrn.py program, but this was quicker.