fix hackPCEOutput function for converToPrn.py
(this same change was previously made to the EmbeddedSampling class)

When outputting tecplot format files, this function would output, on the
first line:

TITLE="the title", VARIABLES=

... with no spaces between TITLE, =, or the first quotes.   As it turns
out, this confuses the convertToPrn.py program in the test suite, and
causes the conversion to fail.  Probably the right thing to do is fix
the convertToPrn.py program, but this was quicker.

This commit also renames two functions:

hackEnsembleOutput   ->  hackPCEOutput
computeEnsembleOutput ->  computePCEOutput