Fixes the KSparse solver for transpose solves
The KSparse interface was not enabled to perform transpose solves.
While the interfaces appeared to support a transpose solve option,
the implementation was not completed to use row and/or column scaling
to perform the transpose solves.

This commit extends the KSparse interface to call the transpose methods
in KSparse and fixes the KSparse solver to use the appropriate scalings
and reoderings when performing the transpose solve.

When running the regression tests there are still failures when using
KSparse due to convergence failures.  Unless time is invested in the
KSparse solver to harden it, this will be the case.

This addresses SON Bug 1207