Improve Support for Special Variable FREQ
This commit improves support for FREQ as a special variable that
returns the current simulation frequency.  It now works correctly
when used on .PRINT lines such as .PRINT AC {FREQ} FREQ

The raw file outputters needed to be modified so that the variable
list still contained the text shown below.  This was done for
backwards compatibility with the existing gold standards in the
regression test suite.  It would be easy to change the variable
name FREQUENCY to FREQ, if needed.

   Variables:
	0	FREQUENCY	frequency

One interface change is that the General External Interface now
looks for the string FREQ rather than the string FREQUENCY.

SON Bug 1212
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1212