Fix syntax error in ceil and floor code
A google groups subscriber pointed out that the xyceBasicTemplates
code emitted an extra closing paren for both the ceil and floor
functions when they are used.  This was due to an extra right paren in
a template call that has never been exercised by any model we've run
through Xyce/ADMS to date.

This was fixed on develop for the sacadoized templates, fixing now on
the bug1168 branch for the non-sacado templates.