Add hacks to explore explicit derivs of analog functions
Qucs and ngspice both do explicit differentiation of analog functions,
but the way they do it actually excludes using their method on
MVSG-HV, and therefore I can't just copy their approach.

The issue is how to handle analog functions that actually have
"output" or "inout" variables.  These are not at all handled well in
either of those simulators, and are not handled particularly well in
Xyce, either.

For starters, ADMS does NOT mark output variables as if they were
assigned to, and does not propagate dependencies to them.  The
adms.implicit.xml templates drill down into expressions and propagate
the dependencies of function arguments into the expression that
contains function calls, and that results in assignments getting the
correct dependencies.  But it does not perform the same operation of
propagating these dependencies into output variables of analog
functions.

This commit takes one tiny baby step in the direction of understanding
how such information can be extracted from the ADMS data tree at the
point of use of a function.  It was not trivial getting here, and in
fact I walked down many blind alleys and wrote many invalid ADMS paths
before finding the way to do it.

Lest I forget, I'm preserving this information in comments to
xyceBasicTemplates in the place where I was able to make ADMS annotate
a function call with the in/out properties of the dummy arguments
corresponding to the real arguments.

I'm also beginning a "NOTES" file that will collect the things I learn
along the way, because the reams of paper and baskets of paper
snowballs I'm creating are simply too easy to lose or forget about.

Finally, I'm starting a file that will hold the templates that
generate the non-sacadoized analog functions needed.

In fact, there is no real reason to generate non-sacadoized analog
functions for the simple "just gimme the value" case --- the templated
version will be fine, and is already mostly correct (modulo weirdness
in FBH that I still don't understand completely).  The real problem
will be generating a function that only returns derivatives.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1168