Allow Y- and Z-Paramter Output in Touchstone Files
This commit adds the capability to have Y-parameters and Z-parameters
output to Touchstone1 and Touchstone2 formatted files for .LIN
analyses.  The outputted parameter format is controlled by a new
Xyce-specific parameter on the .LIN line, where the default value is S:

  .LIN LINTYPE=<S|Y|Z>

The "parameter" field on the "option line" in the header of the
Touchstone file specifies which format (S, Y or Z) was used.

Multiple .LIN lines are allowed in a netlist.  So, both S-, Y- and
Z-parameters can be output from one simulation run, and a single
LIN analysis.

SON Bug 1214
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1214