clean up sacado/kokkos/stokhos includes
This commit is driving by this bit of information that I received from
today:

---
I have recently been made aware of the fact that, as of February 2015,
codes that are using Sacado but not using threading through Kokkos should
be using the Sacado_No_Kokkos.hpp header.  Using the Sacado.hpp header
when threading is not being used is an issue if Xyce is being coupled to
codes that are using Kokkos.  ...

I would suggest changing our headers to Sacado_No_Kokkos, which is fine
for building against Trilinos 12.12.  However, that would mean Xyce could
no longer build against versions of Trilinos before 12.10.
---

The code I have in these files did include Sacado and also other things
such Kokkos_ArithTraits etc.  Most of that was not necessary, and the
code compiles fine if I get rid of most of it.  My use of Stokhos so far
has not relied (directly) on Kokkos at all.  And, at least for the time
being, I am not planning on this.  That may come later.