Allow RF Param Ops to be Used in Expressions
This commit allows the various S(), Y() and Z() operators to be
used in expressions. So, for example, this includes S(), SR(), SI(),
SM(), SP() and SDB().  The fixes for this issue are in the functions
ExpressionInternals::mkfnode_() and ExpressionInternals::tokenize_().

The change to the RFparamsOp class was then needed so that the
S(), Y() and Z() operators, when used in an expression as say
{S(1,1)}, print out (for example) the real-part of S(1,1).  Without
that change, they would print out 2x the correct value.

SON Bug 1199
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1199