Improve Error Messages for Invalid .LIN Lines
This commit makes small changes to the error messages for invalid
DATAFORMAT=<val> and LINTYPE=<val> parameters on a .LIN Line.  The
previous versions did not mention that the error message was generated
by a .LIN line.  It also changed the error message for LINTYPE since
some nominally valid parameter types (e.g., H and G) are valid but
not supported in Xyce.

SON Bug 1199
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1199