Fix typing of instance vars in sensitivity context
Pursuing the L-UTSOI model, two heretofore unnoticed bugs in Xyce/ADMS
were tripped.

The first bug was that the sensitivity code (which was sorta just
slammed in as part of a crisis-mode push) was casually declaring some
classes of variables as Sacado data types without actually checking
that they really needed to be (because they depend on model or
instance parameters).  These variables were all of the sort that ADMS
calls "global_instance" or "global_model" scoped variables.

Later, when the actual code for "evaluteModelEquations",
"evaluateInitialInstance" or "evaluateInitialModel", variables *local*
to those functions *DID* get correctly declared based on whether they
were parameter dependent or not.  The bug was that the earlier
careless declaration of all global_instance real variables as Sacado
types meant that if ever one tried to assign to a
parameter-independent local variable using one of these
global_instance variables on the right-hand side, you'd get a
compilation error.

Fixing this bug requires actually applying the same sort of careful
parameter dependence analysis on global_<model|instance> variables
that one is already performing on all the local variables.

That then exposed a *second* careless bug in the sensitivity code.
All of the sensitivity stuff is basically a variant on the normal
context stuff like "Instance::processParams", "Model::processParams"
or "Instance::updateIntermediateVars", which are all member methods of
their respective classes.  But the sensitivity stuff is all free
functions, with no access to the member data of the class objects.  So
in order to perform the work, the sensitivity operator() methods have
to *copy* the member variables into
appropriately-typed (AdmsSensFadType) local variables and pass them to
the free functions.

It was tacitly being assumed that all such variables were just
doubles.  This breaks if *ever* one has (for whatever reason) an
instance class member that is probe-dependent, and therefore has been
declared in the class as an "AdmsFadType" (the Sacado type used to
house probe-dependent quantities in non-sensitivity context).

It is therefore necessary to test for probe dependence of instance
vars and append a ".val()" as appropriate before copying out the
values.

Note that when these instance vars are getting copied, the copying of
doubles into Sacado variables was always making them have zero
derivatives from Sacado for their derivatives with respect to
parameters, which seems wrong.  But that is actually not a problem,
because what the sensitivity code is really doing in the end is
passing them by reference to the very places where they're getting
recalculated anyway (the
evaluate<InitialInstance|InitialModel|ModelEquations> functions),
which is going to do them all again properly with tracking of
parameter dependence.  In fact, now it may even be pointless (but not
harmful) to be copying the values of the instance variables in the
first place.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1217