added param to exclude sampling stats from output
When running with PCE enabled, you are likely to use a small number of
samples.  For that small number of samples, the sampling code will
compute sampling statistics like mean and std deviation by default.  But
because the sample size is pretty small, this won't be very accurate.

The stats from the PCE appoximation have a better chance of being
accurate.  So, a likely use case is one in which the user wants to
exclude the sample statistics, and include the PCE stats.

Also, for the purposes of testing, this makes it easier, as I can more
conveniently exclude inaccurate numbers from the test.

I have added a new parameter, outputSampleStats, to the ".options
EMBEDDEDSAMPLES" block.  It is "true" by default but can be set to
"false" now.