Stop using function precomputation in sensitivity
One of the big things that has to be done to get non-sacado
derivatives w.r.t. probes (as is done in non-sensitivity context for bug
1168), is to precompute any function calls that appear on the right
hand side of assignments, so that one can also precompute their
derivatives with respect to probes.

This is not in fact necessary in sensitivity context, because we're
using Sacado derivatives to get at derivatives with respect to
parameters.

I had already had a way of telling Xyce/ADMS "don't do function
precomputation here" for various reasons, and it *should* have been
turned on in sensitivity context.  It wasn't, but it was harmless (at
the expense of having more block-local Sacado variables than
necessary).  I'd left it in place in the theory that in the future it
might come in handy if I ever thought to implement derivatives with
respect to both probes and parameters (e.g. do the probes by analytic
explicit differentiation using Sacado variables that did the derivs
for params).  This was a bad idea anyway, because it would really,
really bloat the sensitivity code and make compilation even slower.  I
basically abandoned the thought of doing that, but forgot to disable
the generation.

Now, with the bug 1217 fixes, it is no longer harmless --- the same
sort of bug exists in incorrectly causing variables to be declared as
Sacado that are used as if they're doubles later.

This commit simply turns off this type of code generation in
sensitivity context.  It is so simple I could have just done an
amended commit to my previous, but chose not to because I wanted a
detailed commit message explaining the reason for it.
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1217
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1168