Add PSP103 with self-heating to scripts
The "generate_ADMS.sh" script goes through all Verilog-A
directories for supported devices, rebuilds the C++ with ADMS, and
then compares to the existing C++ code in the src directory.

compare_ADMS.sh just does the compare of all files in the ADMS tree
with those in the src tree.

install_ADMS.sh copies the files out of the ADMS directories into
where they need to be for building.

This commit updates them to build, compare, and install PSP103TVA (the
one with self heating).

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1219