Add Info on SON Bug 1218 to Xyce 6.12 Release Notes
This commit adds a note about SON Bug 1218 to the Xyce 6.12 Release
Notes.  It notes that the minimum required version of Trilinos has
changed.  I'm not sure that this info should actually be in the
Release Notes.  If not then it can be removed later, as part of
the 6.12 Release Process.

This commit also fixed the entry for SON Bug 1210, as part of QA
for that bug.

SON Bug 1218
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1218