Improve Error Checking for RF Param Ops
This commit adds the ability to differentiate between when a .AC
analysis is done versus when a .LIN analysis is done.  This improves
the error checking for the use of the S(), Y() and Z() operators,
and their variants such as SR(), SI(), SM(), SP() and SDB().  The
correct error messages now occurs if this .PRINT AC line is used
in the netlist but a .LIN analysis is not requested.

  .PRINT AC S(1,1) Y(1,1) Z(1,1)

SON Bug 1199
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1199