Add an include of <cstring> where needed
N_IO_CmdParse uses "std::strcpy" and "std::strlen", but does not
include <cstring> itself.  It was somehow mysteriously getting it from
a Sacado header that was indirectly included by some Xyce header it
was including, which must mean that somehow it was getting N_DEV_Pars.h,
which just today got a Sacado include removed.

This led to a compilation failure on RHEL7 with gcc, and possibly others.
The specific error is that strcpy is not in the std:: namespace.  It's
normally declared to be there by the <cstring> header.

I did not trace through the code to figure out what path was taken to
get to N_DEV_Pars.h, because it's not important.  I am inserting
"#include <cstring>" where it is needed, and now it builds on RHEL7
with GCC.