Add <iterator> include to a file
N_TOP_SerialLSUtil.C uses std::back_inserter.  It is supposed to be
the case that this declaration comes from an include of <iterator>.

After the edits to N_DEV_Pars.h today (removing an otherwise
unnecessary Sacado include), lots of things started breaking on
various platforms because headers that they needed were magically
getting satisfied by that Sacado include.  Windows is now failing to
compile N_TOP_SerialLSUtil.C because of this.

Committing and pushing so I can rebuild the failed Jenkins job to see
if this fixes it, since I can't test locally.