remove vestiges of HB solver factory
The PCE solver factory is derived from the ES solver factory, which was
in turn copied from the HB solver factory.  However, a number of
artifacts remained in the PCE factory that were HB specific.  This
commit removes them.

This commit also includes some minor cleanup in the PCE builder and also
the main PCE class.